requirements.txt
Flask==2.0.3
tensorflow==2.10.0
numpy==1.21.2
requests==2.26.0
Pillow==8.4.0
